% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{snake_palettes}
\alias{snake_palettes}
\title{Built-in Color Palettes}
\format{
A named list of 10 character vectors, each with 7 hex color strings.
}
\usage{
snake_palettes
}
\description{
A named list of 10 color palettes for snake plots. Each palette contains
7 anchor colors that can be interpolated to any length with
\code{\link{snake_palette}}.
}
\details{
\describe{
\item{classic}{Diverging red-to-blue. Clean Likert default.}
\item{earth}{Diverging brown-to-teal. Natural, understated.}
\item{ocean}{Diverging coral-to-navy. Warm/cool contrast.}
\item{sunset}{Diverging orange-to-indigo. Vivid but balanced.}
\item{berry}{Diverging rose-to-green. High contrast.}
\item{blues}{Sequential light-to-dark blue.}
\item{greens}{Sequential light-to-dark green.}
\item{grays}{Sequential light-to-dark gray.}
\item{warm}{Sequential cream-to-dark red.}
\item{viridis}{Sequential yellow-green-blue-purple (viridis-inspired).}
}
}
\examples{
snake_palettes$ocean
survey_snake(ema_emotions, colors = snake_palettes$earth,
             tick_shape = "bar", sort_by = "mean")

}
\keyword{datasets}
