% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakeplot-package.R
\docType{package}
\name{snakeplot-package}
\alias{snakeplot}
\alias{snakeplot-package}
\title{snakeplot: Serpentine Plots for Long Timeline, Sequential and Survey Data}
\description{
Visualize long timelines, extended sequences and temporally chained survey responses and experience sampling data using intuitive serpentine (snake) plots. Supports distribution bars, tick-mark plots, inter-item correlation arcs, faceted multi-construct panels, and daily time-of-day positioning for ecological momentary assessment data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mohsaqr/snakeplot}
  \item Report bugs at \url{https://github.com/mohsaqr/snakeplot/issues}
}

}
\author{
\strong{Maintainer}: Mohammed Saqr \email{saqr@saqr.me} [copyright holder]

Authors:
\itemize{
  \item Sonsoles Lopez-Pernas
}

}
\keyword{internal}
