% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakeplot-package.R
\docType{package}
\name{snakeplot-package}
\alias{snakeplot}
\alias{snakeplot-package}
\title{snakeplot: Serpentine (Snake) Plots for Survey and Activity Data}
\description{
Visualize survey responses and experience sampling data using serpentine (snake) layouts where horizontal bands alternate direction row-by-row, connected by U-turn arcs. Supports distribution bars, tick-mark plots, inter-item correlation arcs, faceted multi-construct panels, and daily time-of-day positioning for ecological momentary assessment data. Includes three bundled datasets from Neubauer and Schmiedek (2024) \doi{10.1007/s11618-023-01182-8}. Uses only base 'R' graphics with zero external dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mohsaqr/snakeplot}
  \item Report bugs at \url{https://github.com/mohsaqr/snakeplot/issues}
}

}
\author{
\strong{Maintainer}: Mohammed Saqr \email{saqr@saqr.me} [copyright holder]

Authors:
\itemize{
  \item Sonsoles Lopez-Pernas
}

}
\keyword{internal}
